tests must be moved into benchmark/ directory in order for nosetests to find them
