+++ benchmark/experiments/waterbox_test.py
@@ -16,7 +16,7 @@ from benchmark.experiments.tools import estimate_nonequilibrium_free_energy
 from benchmark.experiments.tools import get_equilibrium_samples, \
     randomization_midpoint_operator, apply_protocol
 from benchmark.integrators.integrators import LangevinSplittingIntegrator
-from code.testsystems import system_params
+from benchmark.testsystems import system_params
 
 matplotlib.use("Agg")
 import matplotlib.pyplot as plt
