+++ benchmark/plotting/__init__.py
@@ -1,3 +1,3 @@
-from .plotting_utilities import generate_figure_filename, plot_scheme_comparison
+from .plotting_utilities import plot, generate_figure_filename, plot_scheme_comparison
 
 __all__ = ["generate_figure_filename", "plot_scheme_comparison"]
