+++ benchmark/testsystems/__init__.py
@@ -3,6 +3,7 @@ from .waterbox import flexible_waterbox, waterbox_constrained
 from .coupled_power_oscillators import coupled_power_oscillators
 from .low_dimensional_systems import quartic, NumbaNonequilibriumSimulator
 from .bookkeepers import EquilibriumSimulator, NonequilibriumSimulator
+from .testsystems import system_params
 
 __all__ = ["alanine_constrained", "alanine_unconstrained", "solvated_alanine_unconstrained",
            "flexible_waterbox", "waterbox_constrained", "coupled_power_oscillators",
