+++ benchmark/utilities/__init__.py
@@ -1,5 +1,5 @@
 from .utils import strip_unit, stderr,\
-    summarize, print_array
+    summarize, print_array, get_summary_string
 
 from .openmm_utilities import keep_only_some_forces,\
     get_total_energy, get_positions, get_velocities,\
